Removed TODOs for removing magento:cache:varnish:ban invocation in deploy procedure
The underlying issue requiring this be used to purge varnish on cache flush (see magento/magento2#3339 for details) was shipped in the Magento 2.0.6 release. However, while cache:clean does send PURGE requests to varnish for the various tags which Magento uses, a cache:flush does not send anything to varnish, requiring a manual PURGE. So this will be left in here to facilitate sending a PURGE to varnish for the entire cache pool.